Stripe/requirements.txt
stripe==7.11.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
web3==6.11.1 